Database Exception handling

+++ src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -3,11 +3,15 @@ package com.devsuperior.dscommerce.services;
 import com.devsuperior.dscommerce.dto.ProductDto;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.services.exceptions.DatabaseException;
 import com.devsuperior.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -38,15 +42,26 @@ public class ProductService {
 
     @Transactional()
     public ProductDto update(Long id, ProductDto productDto) {
-        Product entity = productRepository.getReferenceById(id);
-        copyDtoToEntity(productDto, entity);
-        entity = productRepository.save(entity);
-        return new ProductDto(entity);
+        try {
+            Product entity = productRepository.getReferenceById(id);
+            copyDtoToEntity(productDto, entity);
+            entity = productRepository.save(entity);
+            return new ProductDto(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Product not found");
+        }
     }
 
-    @Transactional()
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        productRepository.deleteById(id);
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Product not found");
+        }
+        try {
+            productRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
